pyproject.toml
[project]
name = "workshop-3"
version = "0.1.0"
description = ""
authors = [
    {name = "kevinMG1601",email = "kevin.andres2636@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "notebook (>=7.4.2,<8.0.0)",
    "ipykernel (>=6.29.5,<7.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "kafka-python (>=2.2.7,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "country-converter (>=1.3,<2.0)",
    "scikit-learn (>=1.6.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
